+++ changelog.md
@@ -15,6 +15,8 @@
 - app failing when powershell is not part of the $PATH enviroment.
 - dock items no updatings paths on store updates.
 - unremovable workspace module on toolbar.
+- missing system tray icons.
+- date not being inmediately updated on settings change.
 
 ## [2.0.14]
 ### hotfix